requirements.txt
pypi-json
netaddr
netifaces
pytest
pytz
PyYAML
requests
websocket-client>=1.2.1
